# src/framework/api/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"exampleResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": true
			},
			"ValidationError": {
				"description": "Represents a validation error",
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"fields": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"fields"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BusinessError": {
				"description": "Represents a business error",
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ProjectSection": {
				"enum": [
					"INVESTMENT",
					"PAYMENTS",
					"FUNDING",
					"MANAGEMENT"
				],
				"type": "string"
			},
			"userCreatedResponse": {
				"properties": {},
				"type": "object",
				"additionalProperties": true
			},
			"IUserCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sta",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "brunodev09"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/example/{agencyId}/{projectType}/agency": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/exampleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"422": {
						"description": "Business Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BusinessError"
								}
							}
						}
					}
				},
				"tags": [
					"example"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "agencyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectType",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ProjectSection"
						}
					}
				]
			}
		},
		"/user/create": {
			"post": {
				"operationId": "Post",
				"responses": {
					"201": {
						"description": "User has been created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/userCreatedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_id": "213b21y3g12gy",
											"name": "customUsername123"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					},
					"422": {
						"description": "Business Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BusinessError"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreationParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}